Refarctored styling
