+++ .vscode/settings.json
@@ -4,4 +4,7 @@
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
+    "[javascript]": {
+        "editor.tabSize": 2
+    }
 }
