+++ client/package.json
@@ -28,6 +28,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-native": "0.76.2",
+    "react-native-actionsheet": "^2.4.2",
     "react-native-chart-kit": "^6.12.0",
     "react-native-element-dropdown": "^2.12.4",
     "react-native-gesture-handler": "~2.20.2",
