+++ client/src/assets/adaptive-icon.png
