+++ client/src/assets/favicon.png
