+++ client/src/components/common/button.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native';
+
+const buttonStyles = StyleSheet.create({
+  button: {
+    allignItems: 'center',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 40,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+    marginVertical: 5, // Space between buttons
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+});
+
+export default buttonStyles;
