+++ client/src/components/common/commonContainer.js
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+const containerStyles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    paddingVertical: 10,
+    paddingHorizontal: 45,
+    top: '-20%',
+    // marginBottom: -150,
+  },
+  switchContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginVertical: 10,
+  },
+});
+
+export default containerStyles;
