+++ client/src/components/common/textInput.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const textInputStyles = StyleSheet.create({
+  textInput: {
+    width: 100,
+    height: 50,
+    backgroundColor: '#fff',
+    borderColor: '#ccc',
+    borderWidth: 1,
+    borderRadius: 10,
+    paddingHorizontal: 10,
+    marginBottom: 15,
+    textAlign: 'center',
+    textAlignVertical: 'center',
+  },
+});
+
+export default textInputStyles;
