+++ client/src/components/styles/Map.styles.js
@@ -1,0 +1,133 @@
+import { StyleSheet, Platform } from 'react-native';
+
+const MapStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    ...(Platform.OS === 'web' ? { height: '100vh' } : {}),
+  },
+  map: {
+    flex: 1,
+    minHeight: 300,
+  },
+  sendButton: {
+    position: 'absolute',
+    bottom: 15,
+    left: '80%',
+    transform: [{ translateX: -50 }],
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    borderRadius: 10,
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    textAlign: 'center',
+  },
+  loginButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '70%',
+    top: '0%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 40,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  PreferencesButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '0%',
+    top: '10%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 40,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+
+  friendScreenButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '70%',
+    top: '50%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 40,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  weatherButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '0%',
+    top: '0%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 40,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  findRouteButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '34%',
+    top: '0%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 35,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  dashboardButton: {
+    position: 'absolute',
+    alignItems: 'center',
+    left: '0%',
+    top: '93%',
+    backgroundColor: '#007bff',
+    paddingVertical: 10,
+    paddingHorizontal: 10,
+    borderRadius: 10,
+    shadowColor: '#000',
+    shadowOpacity: 0.2,
+    shadowOffset: { width: 0, height: 2 },
+    shadowRadius: 5,
+    elevation: 5,
+  },
+  error: {
+    flex: 1,
+    textAlign: 'center',
+    textAlignVertical: 'center',
+    fontSize: 18,
+    color: 'red',
+  },
+  loadingText: {
+    flex: 1,
+    textAlign: 'center',
+    textAlignVertical: 'center',
+    fontSize: 18,
+  },
+});
+
+export default MapStyles;
